requirements.txt
GitPython==3.1.18
python-dotenv==0.17.1